renamed CoinInformation  to Summary
