+++ src/pages/Summary/Summary.js
@@ -15,7 +15,7 @@ import {
   RightLink,
   Subtitle,
   TopPageContent
-} from './CoinInformation.styles';
+} from './Summary.styles';
 
 import {
   CoinPricesData,
@@ -29,14 +29,14 @@ import {
 import getSymbolFromCurrency from 'currency-symbol-map';
 import { Loading, CenterDiv } from 'assets';
 
-class CoinInformation extends Component {
+class Summary extends Component {
   state = {
     isLoading: true,
     hasError: false,
     coinInfo: []
   };
 
-  getCoinInformation = async () => {
+  getSummary = async () => {
     const id = this.props.params.id;
     this.setState({ isLoading: true });
     try {
@@ -64,7 +64,7 @@ class CoinInformation extends Component {
   };
 
   componentDidMount = () => {
-    this.getCoinInformation();
+    this.getSummary();
   };
   render() {
     if (this.state.hasError)
@@ -162,4 +162,4 @@ class CoinInformation extends Component {
   }
 }
 
-export const WrappedCoinInformation = withRouter(CoinInformation);
+export const WrappedSummary = withRouter(Summary);
