+++ src/pages/index.js
@@ -1,3 +1,3 @@
 export { Coins } from './Coins/Coins';
 export { Portfolio } from './Portfolio/Portfolio';
-export { WrappedCoinInformation as CoinInformation } from './CoinInformation/CoinInformation';
+export { WrappedSummary as Summary } from './Summary/Summary';
